+++ Project_2/Utils.h
@@ -52,7 +52,7 @@ public:
 		return sf::Vector2f(0, 0);
 	}
 
-	static float shortestAngle(float angle1, float angle2) {
+	static double shortestAngle(float angle1, float angle2) {
 		return fmod(angle2 - angle1 + 180, 360) - 180;
 	}
 	static int shortestAngle(int angle1, int angle2) {
@@ -62,7 +62,7 @@ public:
 	static int modAngle(int angle) {
 		return (angle + 360) % 360;
 	}
-	static float modAngle(float angle) {
+	static double modAngle(float angle) {
 		return fmod(angle + 360, 360);
 	}
